wgpu: Inline IncompleteDrawType::build
